integrates new features module, plus small refactor
